refactor(starknet_client): move small structs of StateDiff to common (#1381)
