+++ crates/papyrus_common/src/lib.rs
@@ -5,6 +5,7 @@ pub mod block_hash;
 pub mod metrics;
 pub mod patricia_hash_tree;
 pub mod pending_classes;
+pub mod state;
 pub mod transaction_hash;
 
 #[derive(Copy, Clone, Debug, Default, Deserialize, Eq, PartialEq, Serialize)]
