+++ crates/papyrus_common/src/state.rs
@@ -1,0 +1,31 @@
+use serde::{Deserialize, Serialize};
+use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress};
+use starknet_api::hash::StarkFelt;
+use starknet_api::state::StorageKey;
+
+/// A storage entry in a contract.
+#[derive(Debug, Default, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, PartialOrd, Ord)]
+pub struct StorageEntry {
+    pub key: StorageKey,
+    pub value: StarkFelt,
+}
+
+/// A deployed contract in StarkNet.
+#[derive(Debug, Default, Clone, Eq, PartialEq, Hash, Deserialize, Serialize, PartialOrd, Ord)]
+pub struct DeployedContract {
+    pub address: ContractAddress,
+    pub class_hash: ClassHash,
+}
+
+/// A mapping from class hash to the compiled class hash.
+#[derive(Debug, Clone, Default, Eq, PartialEq, Deserialize, Serialize)]
+pub struct DeclaredClassHashEntry {
+    pub class_hash: ClassHash,
+    pub compiled_class_hash: CompiledClassHash,
+}
+
+#[derive(Debug, Clone, Default, Eq, PartialEq, Deserialize, Serialize)]
+pub struct ReplacedClass {
+    pub address: ContractAddress,
+    pub class_hash: ClassHash,
+}
